+++ src/dir_with_dune.ml
@@ -5,7 +5,6 @@ type 'data t =
   ; ctx_dir         : Path.Build.t
   ; data            : 'data
   ; scope           : Scope.t
-  ; kind            : Dune_lang.File_syntax.t
   ; dune_version    : Syntax.Version.t
   }