+++ src/test_rules.boot.ml
@@ -1,4 +1,3 @@
 
-let rules _ ~sctx:_ ~dir:_ ~scope:_ ~expander:_ ~dir_contents:_
-      ~dir_kind:_ =
+let rules _ ~sctx:_ ~dir:_ ~scope:_ ~expander:_ ~dir_contents:_ =
   failwith "test stanza not supported in bootstrap"