+++ test/blackbox-tests/test-cases/dune-jbuild-var-case/run.t
@@ -17,13 +17,13 @@ jbuild files retain the the old names:
   $ dune runtest --root jbuilder-upper
   Entering directory 'jbuilder-upper'
   File "jbuild", line 1, characters 0-0:
-  Warning: jbuild files are deprecated, please convert this file to a dune file
-  instead.
+  Warning: jbuild files are not allowed inside Dune 2.0 projects, please
+  convert this file to a dune file instead.
   Note: You can use "dune upgrade" to convert your project to dune.
 
   $ dune runtest --root jbuilder-upper
   Entering directory 'jbuilder-upper'
   File "jbuild", line 1, characters 0-0:
-  Warning: jbuild files are deprecated, please convert this file to a dune file
-  instead.
+  Warning: jbuild files are not allowed inside Dune 2.0 projects, please
+  convert this file to a dune file instead.
   Note: You can use "dune upgrade" to convert your project to dune.