+++ test/blackbox-tests/test-cases/dup-fields/run.t
@@ -19,8 +19,8 @@ For backward compatibility, it is only a warning in jbuild files:
   $ dune build --root jbuild
   Entering directory 'jbuild'
   File "jbuild", line 1, characters 0-0:
-  Warning: jbuild files are deprecated, please convert this file to a dune file
-  instead.
+  Warning: jbuild files are not allowed inside Dune 2.0 projects, please
+  convert this file to a dune file instead.
   Note: You can use "dune upgrade" to convert your project to dune.
   File "jbuild", line 4, characters 2-21:
   4 |   (action (echo bar))))